# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Razee.io API server.  You can find out more about Razee at [http://razee.io](http://razee.io) or on [slack, #razee](https://ibm-container-service.slack.com/messages/CJQG07QTX).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "2.0.0",
    "title": "Razeedash API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3333",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/api/v2/orgs/": {
      "post": {
        "tags": [
          "orgs"
        ],
        "summary": "Add an organization",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "org-admin-key",
            "description": "The org admin api key",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "name",
            "description": "The org name to be created",
            "required": true,
            "schema": {
              "type":"string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Org"
            }
          },
          "400": {
            "description": "Missing name or org-admin-key header or the org name already exists"
          },
          "401": {
            "description": "invalid org-admin-key"
          }
        }
      },
      "get": {
        "tags": [
          "orgs"
        ],
        "summary": "Get details of an organization or all organizations",
        "description": "To retrive all orgs use: /orgs \n To retrieve a specific org use: /orgs?name=myOrg \n To retrieve a multiple orgs use: /orgs?name=myOrg1&name=myOrg2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "org-admin-key",
            "description": "organization admin key",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "description": "The org name(s) to search. ex: /orgs?name=myOrg1",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OrgArray"
            }
          },
          "400": {
            "description": "Missing name or org-admin-key header or the org name already exists"
          },
          "401": {
            "description": "invalid org-admin-key"
          },
          "500": {
            "description": "Error searching for orgs"
          }
        }
      }
    },
...
    "/api/v2/clusters/": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Get clusters",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "razee-org-key",
            "description": "organization api key",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "org-admin-key",
            "description": "organization admin key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cluster"
            }
          },
          "400": {
            "description": "Missing org-admin-key header"
          },
          "401": {
            "description": "Invalid org-admin-key"
          }
        }
      }
    },
...
    "/api/v2/resources/": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get a resource definition",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "org-admin-key",
            "description": "The org admin api key",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "kind",
            "type": "string",
            "description": "The resource type to search. ex: /resources?kind=Deployment",
            "required": false
          },
          {
            "in": "query",
            "name": "name",
            "type": "string",
            "description": "The resource name to search. ex: /resources?name=myResource",
            "required": false
          },
          {
            "in": "query",
            "name": "skip",
            "type": "string",
            "description": "Skip the first 25 results. ex: /resources?skip=25",
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "type": "string",
            "description": "Limit the number of resources to return.  Max limit is 10000 and the default is 25. ex: /resources?limit=100",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Resource"
            }
          },
          "400": {
            "description": "Missing name or org-admin-key header or the org name already exists"
          },
          "401": {
            "description": "invalid org-admin-key"
          }
        }
      }
    },
    "/channels": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Adds a new channel",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Gets all channels for an org",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
  "securityDefinitions": {
    "razee-org-key": {
      "type": "apiKey",
      "name": "razee-org-key",
      "in": "header"
    },
    "org-admin-key": {
      "type": "apiKey",
      "name": "org-admin-key",
      "in": "header"
    }
  },
  "definitions": {
    "Org": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "orgKeys": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "created": {
          "type": "string"
        },
        "updated": {
          "type": "string"
        }
      }
    },
...
    "OrgArray": {
      "type": "array",
      "items": {
        "type": "object",
        "$ref": "#/definitions/Org"
      }
    },
    "Cluster": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "cluster_id": {
          "type": "string"
        },
        "org_id": {
          "type": "string"
        },
        "metadata": {
          "type": "object"
        }
      }
    },
...
    "Resource": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "SYNC",
            "POLLED",
            "MODIFIED",
            "ADDED",
            "DELETED"
          ]
        },
        "data": {
          "type": "object",
          "properties": {
            "kind": {
              "type": "string"
            },
            "apiVersion": {
              "type": "string"
            },
            "spec": {
              "type": "object"
            }
          }
        },
        "cluster_id": { "type": "string" },
        "org_id": { "type": "string" },
        "selfLink": { "type": "string" },
        "created": { "type": "string", "format": "date" },
        "updated": { "type": "string", "format": "date" },
        "deleted": { "type": "boolean" },
        "hash": { "type": "string" },
        "searchableData": {
          "type": "object",
          "properties": {
            "kind": { "type": "string" },
            "name": { "type": "string" },
            "namespace": { "type": "string" },
            "apiVersion": { "type": "string" }
          }
        },
        "searchableDataHash": { "type": "string" },
        "searchableDataHist": {
          "type": "object",
          "properties": {
            "kind": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "timestamp": { "type": "integer"},
                  "val": { "type": "string"}
                }
              }
            },
            "name": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "timestamp": { "type": "integer"},
                  "val": { "type": "string"}
                }
              }
            },
            "namespace": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "timestamp": { "type": "integer"},
                  "val": { "type": "string"}
                }
              }
            },
            "apiVersion": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "timestamp": { "type": "integer"},
                  "val": { "type": "string"}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}